ENH: jammed in an ipython console
Quick hack in, needs to be cleaned up a bit. Reused the embedded ipython
console file from the pyDAT project I started hat never went anywhere.

Can access the edqplot app object from within the ipython console by
using the 'edqplot_app' variable that already is in lace in the console.

The app object itself is a mess, so interacting with it right now is no
fun, but it could be. ;)

for example, to move the zoom region to a different start/stop time do

edqplot_app._file_data_plot.setRegionBounds(start_time, end_time)

and, as long as the 2 times are within the time series, the region will
move to where requested and all stats get updated based on the new
region location.